Added Authentication-method-specific sub-flow aware of and enforcing matching of Principal
Where a specific authentication method is required, CAS first requires the basic username-and-password login. The result of that basic login is CAS knows about an authenticated Principal.

Here's the punchline: sub-flows modeling particular authentication methods need to be aware of, reflect, and enforce that the additional authentication they are obtaining authenticates that same Principal.