Issue#59
This design and implementation uses ArgumentExtractor machinery to stack
up the mfa method source and creatiion of an appropreate mfa service
based on that. The outer extractor looks for registered service attribute
definition [auth_method] as its source and if supported method is found
uses that to create an mfa service. If not found, ONLY THEN delegates
to a wrapped extractor implementation that uses HTTP request parameter
[authn_method] as its source.

This way, only ONE mfa authn method definition source could be active
at a time which would then result in creation of an appropreate mfa service
instance that is used by the rest of cas-mfa machinery to perform its
mfa authentication function.